Created and Modified Functions, Plots, and Data
Modified functions to extract data from the nba_api and store the data
in corresponding csv files. Wrote a function for halloffamers that
sums the hall of famers from the Eastern and Western conferences.
Add additional functions for each stat that creates plots and data
visualizations for the csv files. Deleted empty files to clean up the
repo.